Remove polysilicon that is part of capacitor
This commit adds support for a new "Capacitors" layer.
Polygons in this layer should fully overlap the bits of polysilicon that
are part of a capacitor.

Capacitors are where polysilicon and diffusion overlap, but there
is no gate underneath.

This is mainly cosmetic to avoid spurious warnings about gates
not having two electrodes.